more docs, more logic in render
